Chamba de hoy.
1. Se retrasó lo del cálculo del flujo de masa debido a que primero se
deben de calcular los coeficientes agrupados para la ecuación de
momentum.
2. Se comenzó el cálculo de la conductancia difusiva, agregando los
archivos `conductancia_difusiva.*`.

+++ src/malla_por_bloques.cpp
@@ -315,7 +315,7 @@ namespace Malla {
     );
 
     // Nodos que deben de ser borrados en y
-    for (int i=0; i<coord_final_y.size()-1; ++i) {
+    for (int i=0; i<(int)coord_final_y.size()-1; ++i) {
       y_tmp.erase
       (
         std::remove(y_tmp.begin(),y_tmp.end(),coord_final_y[i]),y_tmp.end()
